741 op amp model (#330)

+++ src/com/lushprojects/circuitjs1/client/TransistorElm.java
@@ -325,7 +325,7 @@ package com.lushprojects.circuitjs1.client;
 	    }
 	}
 	
-	void setBeta(float b) {
+	void setBeta(double b) {
 	    beta = b;
 	    setup();
 	}
